requirements.txt
discord.py>=2.3.0
qbittorrent-api>=2023.10.50
python-dotenv>=1.0.0

